+++ src/vcsdaemon.nim
@@ -17,7 +17,7 @@ const ETIMER = -4
 const ECONN = -5
 
 const STATIC_HELP_TEXT = static_read("../help.txt")
-let HELP_TEXT = "Svndaemon v" & VERSION_STR & "\n\n" & STATIC_HELP_TEXT
+let HELP_TEXT = "vcsdaemon v" & VERSION_STR & "\n\n" & STATIC_HELP_TEXT
 
 var do_exit = false
 
@@ -90,7 +90,7 @@ if timer_create(CLOCK_REALTIME, nil, timer) < 0:
 var tspec: Itimerspec
 tspec.it_value = Timespec(tv_sec: posix.Time(10), tv_nsec: 0)
 
-log.info("Svndaemon started $1.", utc(times.now()))
+log.info("vcsdaemon started $1.", utc(times.now()))
 
 # Main program loop.
 var trackers: seq[Tracker]